Split submit-requests.test.ts into two files
